+++ resources/assets/components/App.js
@@ -12,7 +12,7 @@ import FeedContainer from '../containers/FeedContainer';
 import ContentPageContainer from '../containers/ContentPageContainer';
 import NotFound from './NotFound';
 
-import observe from '../analytics';
+import observe from '../middleware/analytics';
 
 // Set the application "base name" to /campaigns/:slug so all pages are relative to that.
 const basename = window.location.pathname.split('/').slice(0, 3).join('/');
